# docs/openapi/season-service-v1.yaml
openapi: 3.1.0
info:
  title: ACM Season Service API
  description: API documentation for ACM Season Service
  license:
    name: Apache 2.0
    url: https://springdoc.org
  version: 1.0.0
servers:
- url: http://localhost
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/disease-treatments/{id}:
    put:
      tags:
      - disease-record-controller
      operationId: updateTreatment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDiseaseTreatmentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseDiseaseTreatmentResponse"
    delete:
      tags:
      - disease-record-controller
      operationId: deleteTreatment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
  /api/v1/disease-records/{id}:
    get:
      tags:
      - disease-record-controller
      operationId: getDiseaseRecordDetail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseDiseaseRecordDetailResponse"
    put:
      tags:
      - disease-record-controller
      operationId: updateDiseaseRecord
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDiseaseRecordRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseDiseaseRecordResponse"
    delete:
      tags:
      - disease-record-controller
      operationId: deleteDiseaseRecord
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /api/v1/seasons/{seasonId}/disease-records:
    get:
      tags:
      - disease-record-controller
      operationId: listDiseaseRecordsBySeason
      parameters:
      - name: seasonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: severity
        in: query
        required: false
        schema:
          type: string
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: fromDetectedAt
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: toDetectedAt
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseDiseaseRecordResponse"
    post:
      tags:
      - disease-record-controller
      operationId: createDiseaseRecord
      parameters:
      - name: seasonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiseaseRecordRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseDiseaseRecordResponse"
...
  /api/v1/disease-records/{id}/treatments:
    get:
      tags:
      - disease-record-controller
      operationId: listTreatments
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseDiseaseTreatmentResponse"
    post:
      tags:
      - disease-record-controller
      operationId: createTreatment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiseaseTreatmentRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseDiseaseTreatmentResponse"
...
components:
  schemas:
...
    UpdateDiseaseTreatmentRequest:
      type: object
      properties:
        treatedAt:
          type: string
          format: date-time
        method:
          type: string
          maxLength: 100
          minLength: 0
        supplyItemId:
          type: integer
          format: int32
        supplyLotId:
          type: integer
          format: int32
        materialName:
          type: string
          maxLength: 150
          minLength: 0
        quantityUsed:
          type: number
          minimum: 0.0
        unit:
          type: string
          maxLength: 20
          minLength: 0
        costAmount:
          type: number
          minimum: 0.0
        expenseId:
          type: integer
          format: int32
        effectiveness:
          type: string
        resultSummary:
          type: string
          maxLength: 4000
          minLength: 0
        nextReviewAt:
          type: string
          format: date
        notes:
          type: string
          maxLength: 4000
          minLength: 0
    ApiResponseDiseaseTreatmentResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/DiseaseTreatmentResponse"
        data:
          $ref: "#/components/schemas/DiseaseTreatmentResponse"
        generatedAt:
          type: string
    DiseaseTreatmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        diseaseRecordId:
          type: integer
          format: int32
        treatedAt:
          type: string
          format: date-time
        method:
          type: string
        supplyItemId:
          type: integer
          format: int32
        supplyItemName:
          type: string
        supplyLotId:
          type: integer
          format: int32
        batchCode:
          type: string
        materialName:
          type: string
        quantityUsed:
          type: number
        unit:
          type: string
        costAmount:
          type: number
        expenseId:
          type: integer
          format: int32
        effectiveness:
          type: string
        resultSummary:
          type: string
        nextReviewAt:
          type: string
          format: date
        notes:
          type: string
        createdByUserId:
          type: integer
          format: int64
        createdByUsername:
          type: string
        createdByDisplayName:
          type: string
        createdByType:
          type: string
        canEdit:
          type: boolean
        canDelete:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateDiseaseRecordRequest:
      type: object
      properties:
        diseaseName:
          type: string
          maxLength: 150
          minLength: 0
        symptomSummary:
          type: string
          maxLength: 4000
          minLength: 0
        severity:
          type: string
        status:
          type: string
          maxLength: 30
          minLength: 0
        detectedAt:
          type: string
          format: date-time
        affectedPlantCount:
          type: integer
          format: int32
          minimum: 0
        affectedAreaValue:
          type: number
          minimum: 0.0
        affectedAreaUnit:
          type: string
          maxLength: 20
          minLength: 0
        evidenceUrl:
          type: string
          maxLength: 1000
          minLength: 0
        notes:
          type: string
          maxLength: 4000
          minLength: 0
        incidentId:
          type: integer
          format: int32
    ApiResponseDiseaseRecordResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/DiseaseRecordResponse"
        data:
          $ref: "#/components/schemas/DiseaseRecordResponse"
        generatedAt:
          type: string
    DiseaseRecordResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        seasonId:
          type: integer
          format: int32
        seasonName:
          type: string
        plotId:
          type: integer
          format: int32
        plotName:
          type: string
        cropId:
          type: integer
          format: int32
        cropName:
          type: string
        varietyId:
          type: integer
          format: int32
        varietyName:
          type: string
        reportedByUserId:
          type: integer
          format: int64
        reportedByUsername:
          type: string
        reportedByDisplayName:
          type: string
        reportedByType:
          type: string
        canEdit:
          type: boolean
        canDelete:
          type: boolean
        incidentId:
          type: integer
          format: int32
        diseaseName:
          type: string
        symptomSummary:
          type: string
        severity:
          type: string
        status:
          type: string
        detectedAt:
          type: string
          format: date-time
        affectedPlantCount:
          type: integer
          format: int32
        affectedAreaValue:
          type: number
        affectedAreaUnit:
          type: string
        evidenceUrl:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        treatmentCount:
          type: integer
          format: int64
        latestTreatmentAt:
          type: string
          format: date-time
...
    CreateDiseaseRecordRequest:
      type: object
      properties:
        diseaseName:
          type: string
          maxLength: 150
          minLength: 0
        symptomSummary:
          type: string
          maxLength: 4000
          minLength: 0
        severity:
          type: string
        status:
          type: string
          maxLength: 30
          minLength: 0
        detectedAt:
          type: string
          format: date-time
        affectedPlantCount:
          type: integer
          format: int32
          minimum: 0
        affectedAreaValue:
          type: number
          minimum: 0.0
        affectedAreaUnit:
          type: string
          maxLength: 20
          minLength: 0
        evidenceUrl:
          type: string
          maxLength: 1000
          minLength: 0
        notes:
          type: string
          maxLength: 4000
          minLength: 0
        incidentId:
          type: integer
          format: int32
      required:
      - detectedAt
      - diseaseName
      - severity
...
    CreateDiseaseTreatmentRequest:
      type: object
      properties:
        treatedAt:
          type: string
          format: date-time
        method:
          type: string
          maxLength: 100
          minLength: 0
        supplyItemId:
          type: integer
          format: int32
        supplyLotId:
          type: integer
          format: int32
        materialName:
          type: string
          maxLength: 150
          minLength: 0
        quantityUsed:
          type: number
          minimum: 0.0
        unit:
          type: string
          maxLength: 20
          minLength: 0
        costAmount:
          type: number
          minimum: 0.0
        expenseId:
          type: integer
          format: int32
        effectiveness:
          type: string
        resultSummary:
          type: string
          maxLength: 4000
          minLength: 0
        nextReviewAt:
          type: string
          format: date
        notes:
          type: string
          maxLength: 4000
          minLength: 0
      required:
      - method
      - treatedAt
...
    ApiResponsePageResponseDiseaseRecordResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/PageResponseDiseaseRecordResponse"
        data:
          $ref: "#/components/schemas/PageResponseDiseaseRecordResponse"
        generatedAt:
          type: string
    PageResponseDiseaseRecordResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DiseaseRecordResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ApiResponseDiseaseRecordDetailResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/DiseaseRecordDetailResponse"
        data:
          $ref: "#/components/schemas/DiseaseRecordDetailResponse"
        generatedAt:
          type: string
    DiseaseRecordDetailResponse:
      type: object
      properties:
        record:
          $ref: "#/components/schemas/DiseaseRecordResponse"
        treatments:
          type: array
          items:
            $ref: "#/components/schemas/DiseaseTreatmentResponse"
        treatmentCount:
          type: integer
          format: int64
        latestTreatmentAt:
          type: string
          format: date-time
        totalTreatmentCost:
          type: number
    ApiResponsePageResponseDiseaseTreatmentResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/PageResponseDiseaseTreatmentResponse"
        data:
          $ref: "#/components/schemas/PageResponseDiseaseTreatmentResponse"
        generatedAt:
          type: string
    PageResponseDiseaseTreatmentResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DiseaseTreatmentResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    ApiResponseVoid:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          type: object
        data:
          type: object
        generatedAt:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT